Move stories and tests to stories and test directories
