+++ packages/components/src/calendar/stories/date-calendar.story.tsx
@@ -28,12 +28,12 @@ import { useState, useEffect } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { DateRangeCalendar, TZDate } from '../../';
+import { DateCalendar, TZDate } from '..';
 import type { Meta, StoryObj } from '@storybook/react';
 
-const meta: Meta< typeof DateRangeCalendar > = {
-	title: 'UI/DateRangeCalendar',
-	component: DateRangeCalendar,
+const meta: Meta< typeof DateCalendar > = {
+	title: 'UI/DateCalendar',
+	component: DateCalendar,
 	argTypes: {
 		locale: {
 			options: [
@@ -89,8 +89,8 @@ const meta: Meta< typeof DateRangeCalendar > = {
 		labels: {
 			control: false,
 		},
-		defaultSelected: { control: false },
-		selected: { control: false },
+		defaultSelected: { control: 'date' },
+		selected: { control: 'date' },
 		onSelect: {
 			control: false,
 		},
@@ -109,7 +109,7 @@ const meta: Meta< typeof DateRangeCalendar > = {
 };
 export default meta;
 
-type Story = StoryObj< typeof DateRangeCalendar >;
+type Story = StoryObj< typeof DateCalendar >;
 
 export const Default: Story = {};
 
@@ -150,13 +150,9 @@ const nextMonthYear =
 		? new Date().getFullYear() + 1
 		: new Date().getFullYear();
 const firstDayOfNextMonth = new Date( nextMonthYear, nextMonth, 1 );
-const fourthDayOfNextMonth = new Date( nextMonthYear, nextMonth, 4 );
-export const WithSelectedRangeAndMonth: Story = {
+export const WithSelectedDateAndMonth: Story = {
 	args: {
-		defaultSelected: {
-			from: firstDayOfNextMonth,
-			to: fourthDayOfNextMonth,
-		},
+		defaultSelected: firstDayOfNextMonth,
 		defaultMonth: firstDayOfNextMonth,
 	},
 };
@@ -166,40 +162,28 @@ export const WithSelectedRangeAndMonth: Story = {
  */
 export const WithTimeZone: Story = {
 	render: function DateCalendarWithTimeZone( args ) {
-		const [ range, setRange ] = useState< typeof args.selected | null >(
-			null
-		);
+		const [ selected, setSelected ] = useState< TZDate | null >( null );
 
 		useEffect( () => {
-			setRange(
-				// Select from one week from today to two weeks from today
-				// every time the timezone changes.
-				{
-					from: new TZDate(
-						new Date().setDate( new Date().getDate() + 7 ),
-						args.timeZone
-					),
-					to: new TZDate(
-						new Date().setDate( new Date().getDate() + 14 ),
-						args.timeZone
-					),
-				}
+			setSelected(
+				// Select one week from today every time the time zone changes.
+				new TZDate(
+					new Date().setDate( new Date().getDate() + 7 ),
+					args.timeZone
+				)
 			);
 		}, [ args.timeZone ] );
 
 		return (
 			<>
-				<DateRangeCalendar
+				<DateCalendar
 					{ ...args }
-					selected={ range }
+					selected={ selected }
 					onSelect={ ( selectedDate, ...rest ) => {
-						setRange(
-							// Set controlled state to null if there's no selection
-							! selectedDate ||
-								( selectedDate.from === undefined &&
-									selectedDate.to === undefined )
-								? null
-								: selectedDate
+						setSelected(
+							selectedDate
+								? new TZDate( selectedDate, args.timeZone )
+								: null
 						);
 						args.onSelect?.( selectedDate, ...rest );
 					} }
@@ -210,11 +194,11 @@ export const WithTimeZone: Story = {
 						},
 					] }
 				/>
+
 				<p>
 					Calendar set to { args.timeZone ?? 'current' } timezone,
 					disabling selection for all dates before today, and starting
-					with a default date range of 1 week from today to 2 weeks
-					from today.
+					with a default date of 1 week from today.
 				</p>
 			</>
 		);
@@ -223,6 +207,12 @@ export const WithTimeZone: Story = {
 		timeZone: 'Pacific/Auckland',
 	},
 	argTypes: {
+		selected: {
+			control: false,
+		},
+		defaultSelected: {
+			control: false,
+		},
 		disabled: {
 			control: false,
 		},
